requirements.txt
beautifulsoup4==4.9.3
cachelib==0.4.1
certifi==2021.10.8
chardet==4.0.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Session==0.4.0
Flask-WTF==1.0.0
googlesearch-python==1.0.1
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
nltk==3.6.5
numpy==1.21.4
pdfminer==20191125
pycryptodome==3.12.0
regex==2021.11.10
requests==2.25.1
scikit-learn==0.24.1
scipy==1.7.3
soupsieve==2.3.1
threadpoolctl==3.0.0
tqdm==4.62.3
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==2.3.3
